\name{HPCE}
\alias{HPCE}
\title{Health Care price index for the United States.}
\usage{HPCE}
\description{Health Care price index for the United States.}
\references{\emph{www.bea.gov} \cr
\tabular{ll}{
\bold{Table 2.5.4} \tab \cr
\bold{Area:}                   \tab U.S. \cr
\bold{Item:}                   \tab Health \cr
\bold{Base Period:}            \tab 2012=100 \cr
}}
\docType{data}
