\name{PCE}
\alias{PCE}
\title{Personal Consumption Expenditure (PCE) for the United States.}
\usage{PCE}
\description{Personal Consumption Expenditure (PCE) for the United States.}
\references{\emph{www.bea.gov} \cr
\tabular{ll}{
\bold{table 2.5.4} \tab \cr
\bold{Area:}                   \tab U.S. \cr
\bold{Item:}                   \tab All items \cr
\bold{Base Period:}            \tab 2012=100 \cr
}}
\docType{data}
